Updated openstack/openstack Project: openstack/python-keystoneclient  6c1a9b49dc2ffcaa743ebf866c7281622ffad8ad
get_endpoint should return the override

If your adapter has an endpoint_override set then this value will be
consumed by session and used in preference to whatever you give to
endpoint_filter.

This means that if you ask the adapter for the endpoint it is going to
use to query a URL you expect to get back the override because this is
where it will be sent.

Closes-Bug: #1400174
Change-Id: I707e549a4fa349d0e9a0bdac61a2573aa2e5b434