Updated openstack/openstack Project: openstack/python-keystoneclient  10860db5f155052ca0d353ce9058e1c16eec8437
Take plugin params from ENV rather than default

The way the argparse options were being structured, if there was a
default value set on the option it would use this value as the default
and not check the environment variables.

This is wrong, we expect the environment variables to be used and the
default value to be the final fallback.

Change-Id: Ifbd68c9de329c2e0c70824ba873caa579e8e86d0
Closes-Bug: #1388076