Updated openstack/openstack Project: openstack/cinder  a73b1cc349e30de4d90f077e73eb705040161dfb
Isolate Cinder Attach and Connect in Base Driver

The base Cinder Driver combines the Attach process and the
actual Connect process all under the Attach method.  This makes
sense, but the method can be split between the parts that perform
things like DB updates and RPC calls and those that just take iqn
info and actually make connections to the Cinder Volume Node.

This would allow Drivers to utilize the connector portion of the
code to add enhancements of their own.

Change-Id: I9e3cd24dc978e7c7672bf1eb09626428d2ba144d