Updated openstack/openstack Project: openstack/tempest  c7b08af36f140c649dbda6ad17e536f1f3762bda
Change neutron client methods to return one value and update tests

Also removed redundant ok checks in javelin.py and put in a work-around
for the existing assumption that all client 'list_' methods have the same
signature.

It was necessary to change the mocked network calls in the unit tests.
Since the mocked status is never referenced, it was just removed.

Partially implements: blueprint clients-return-one-value

Change-Id: I76de07e984f13e3eec4977a4d823d1ef367e8f46