Updated openstack/openstack Project: openstack/heat  4fc8985f291b3889ef057655b175a277ba19545d
Add PropertyUnspecifiedError exception

If we have situation, when we must specify at
least one of two conflicted resources, and there
is no specified resource, we get StackValidationFailed
with error msg looks like 'Either resource_1 or
resource_2 must be specified'. There are a few places,
where this situation appears. So it will be reasonable
to add special exception for these situations.

Change-Id: Ib107eff4fe8030fdbb40aa267b9e2b394ff7a186