Updated openstack/openstack Project: openstack/horizon  5315911e148a15396c048581f9da3511a9176a67
Add volume type encryption delete

This patch adds support to horizon for volume type encryption
delete.  The contents of this patch were originally a part of
a combined volume type encryption update and delete patch,
(https://review.openstack.org/#/c/72024/) but the delete
portion has been moved to this patch since the pieces do not
depend on each other.

The modifications are made to the Admin Volume Type table, and
add the Delete Encryption action to the action column. Unit
tests are also included.

Change-Id: Ide9109ca22a04b9121bc78eeea26be9ef2adbf97
Implements: blueprint integration-with-cinder-volume-encryption