Updated openstack/openstack Project: openstack/oslo.vmware  93740df5d5c6e636758feab840d8fcdd0efe0ab0
Allow checking api session

session._is_current_session_active should be made available to
consumers of the library as a supported api. It will be very
helpful to know the state of the session.
E.g glance vsphere store bug 1402354.

I verified that the old private method is not in use by other
openstack projects.

Change-Id: I1c244c7b0edbb50e5a84e6e6f6081b20c70451fb