Updated openstack/openstack Project: openstack/neutron  ef5d4f8bee25024b688d1dd00cf5a689f0943500
ml2: remove superfluous %s in LOG.debug() format

ml2.db.get_dynamic_segment() includes this line:

   LOG.debug("No dynamic segment %s found for "
	     "Network:%(network_id)s, "
	     "Physical network:%(physnet)s, "
	     "segmentation_id:%(segmentation_id)s",
	     {'network_id': network_id,
	      'physnet': physical_network,
	      'segmentation_id': segmentation_id})

Note the superfluous %s in the format string.  At run-time, %s prints
the args hash again and doesn't cause an error, but this is clearly
unintended.

Since there doesn't seem to be any value that was meant to be used
instead, this change simply removes the %s.

Change-Id: I414c7b4497f59927d8b6ec5f057dca19aa4ef122
Closes-Bug: #1404782