Updated openstack/openstack Project: openstack/tempest  7ace5a0da46f355635ba21de01593526151c5174
Fix dhcpv6-stateful tempest test to validate only valid use-case

According to IPv6 Radvd implementation[1], when ipv6_ra_mode is not set
and ipv6_address_mode is set to dhcpv6-stateful, Neutron would be using
an external router for Router Advertisements. Such subnets are not
supposed to be associated with Neutron Router as they are meant to be
used with external router.

A recent commit I0a063e543e320ea625a5411547bce7fa2ad66b7d is causing
failure of the Neutron patch[2] in gate. This patch modifies the tempest
code to remove this invalid use-case and validate only the valid
use-cases.

[1] - Neutron ipv6-radvd-ra blueprint
[2] - https://review.openstack.org/#/c/136733/6

Closes-Bug: #1404139
Change-Id: Ib17d1e92d5491aa49a58717ba158f339a1c5f366