Updated openstack/openstack Project: openstack/nova  4990e60f2dc43f794d64d26c1834da18170d2eea
libvirt: introduce new helper for getting libvirt domain

Introduce a get_domain() method to obtain a libvirt domain
from a Nova instance object. This slightly simplifies the
callers, avoiding the need for them to know that the lookup
is done based on the name (as opposed to ID or UUID).

Blueprint: libvirt-driver-class-refactor
Change-Id: Iea29950f3991a8051f48263f43c7b9b861d9e640