Updated openstack/openstack Project: openstack/tempest  840eafaee94c88310595b5a0a4b2277295977070
Merge network clients to a single class

There are two network clients as a base one and a JSON one. This
implementation way is different from the other projects' clients.
Now we need to implement a rest client method as a library function,
and it is necessary to make the rest implementations consistent now.
Then, this patch merges network clients to a single class like the
other projects' ones.

Change-Id: I8373f372dcc7b93c185bb5fc67f6a84429505672