Updated openstack/openstack Project: openstack/tempest  0cd316bd60c00d1a49e1538605fc7344dfb00729
Separate build_interval/timeout from RestClient

RestClient used compute build_interval/timeout values as the default,
even if the other projects' clients. This implementation is not useful
when RestClient is changed to a library class.
Then this patch separates compute build_interval/timeout values from
the class and set the same static values as the default.

Change-Id: Ibc4307b406b71db258975da90a39c1f51ad27e12