Updated openstack/openstack Project: openstack/cinder  c89875d1dcf0d349631605fe2303d4e261703c4d
Replace the NetApp driver proxy layer with a proper factory.

The proxy implementation required a hack in the Abstract Base
Classes project in Cinder core.  This factory implementation
removes the need for the hack.  It also improves logging, as
the actual driver object is provided to the volume manager
instead of the driver proxy.

Change-Id: I28f4798b9e799547d1696d2f135d3302b3c553fe
Closes-Bug: 1405060