Updated openstack/openstack Project: openstack/cinder  e2f9ec4069f86e083d700285aad8c809570d04d1
Expand the description of extra_capabilities

The existing description of extra_capabilities was a bit terse,
meaning deployers couldn't easily gauge the power of this option.
This patch expands the textual description using information from
the blueprint and original commit message, so it can be picked
up automatically by the documentation.

Change-Id: Iac2b6d8a3ce4875a8c1d9874dd7f0e4cf7c435c0
Closes-Bug: 1288044