Updated openstack/openstack Project: openstack/cinder  4efc82cc8dd1039c9089e9f68b35b904f7f7d179
Handle the volume not found case in the VMAX driver

When a volume cannot be found on the array, VMAX driver throws
an exception. This brings a volume into error_deleting state during
the volume deletion operation. This patch fixed it by returning None
when a volume cannot be found.

Closes-Bug: #1395830
Change-Id: Iad714ff6e577bc9d7cf69b8654a7eb726969279e