Updated openstack/openstack Project: openstack/cinder  17c2f6857bf9cdad972238695b415b327a9e25d9
Sync the latest loopingcall module from oslo-incubator

We have not done a sync of loopingcall since its dependency
upon the old i18n module was removed.  To get that old code
removed we need to do this sync.

Current HEAD in OSLO:
---------------------
commit 36b0e8570b449129d6d474c03b02ceb62edb78df
Date:   Thu Dec 11 11:27:08 2014 +0100
We shouldn't replace `oslo-incubator` in comments

Change being merged with this patch:
---------------------
5d40e143 -  Remove code that moved to oslo.i18n

Change-Id: I2fd9cb73cf347adfb2ce203e45d5059604ed5f7a