Updated openstack/openstack Project: openstack/neutron  06b2e816ac6a8fab3b7fa3d8a148c753c690a49b
Tighten dnsmasq version regex

The previous regex forgot to escape the '.', resulting a more liberal
match than intended.  Luckily it continued to work, since the dnsmasq
version is the first number that appears in the --version output.

This change improves the regex to correctly escape the '.' (as
presumably originally intended) and to look for the prefix "version ".

Change-Id: I12548a7b2c21262aa54c84be4e31bda15eab8d52