Updated openstack/openstack Project: openstack/cinder  4c237854e7a3473fa07acf0578571fb3b430c47b
Brick: fix bug in tgt conf for volume

872b7ca0fc5ce4911fdaadee464d46be5a38058d was intended to explicitly
specify the driver being used by tgt for each target. The issue is that
'lld' is not a valid directive therefore the default is being used. The
default used by tgt is iscsi.

This commit replaces 'lld' with the correct directive of 'driver'.

Change-Id: I6153ee8b24039aedc4ab471f7d00517411791dac
Closes-bug: #1396265