Updated openstack/openstack Project: openstack/tempest  94c593d3d14083561d629f30c23b5272aa61bef0
Merge telemetry clients to a single class

There are two telemetry clients as a base one and a JSON one. This
implementation way is different from the other projects' clients.
Now we need to implement a rest client method as a library function,
and it is necessary to make the rest implementations consistent now.
Then, this patch merges telemetry clients to a single class like the
other projects' ones.

Change-Id: Ifbe058df9c4e269a359a65f31efd1d34f208d253