Updated openstack/openstack Project: openstack/nova  56c778198a49ae4bcbc5ff931ed64b201a31ff22
Xenapi: Allow volume backed instances to migrate

The migrate process had no provisions for a non image backed instance.
This remedies that by doing a few things:

Don't snapshot the root disk on the source host.
Don't try to import the root disk on the destination host.
Connect the volume to the destination host as the boot device.

Change-Id: Ia453cf33a36803b1b5447c6240988a3dd603f328
Closes-bug: #1375480