Updated openstack/openstack Project: openstack/heat  12caa51b98e9c61fb0c55ec390d78e2c9a23c618
Added pseudo param OS::project_id

project id can be used to identify or tag which project the resources
belong to. The project id can be found in cfn pseudo param AWS::StackId, but when we
switch to hot, OS::stack_id is a pure stack id, no project info in it. So
we add OS::project_id to meet the similar scenario requirements.

Change-Id: I3db9db7c59cfa0ff3005a97cd0cb7962928598ab