Updated openstack/openstack Project: openstack/horizon  3a492dd1eafc6b87503c15e02f6a0228c4804659
Potential 500 pages on allowed methods removed

The VPN, Firewall and Loadbalancers panels all have checks to see if the
appropriate service is enabled. This call requires a neutron API call
that is not protected in a try/except block. If there is a problem with
the neutron service, on login, the user encounters the 500 page and is
unable to do anything in Horizon.

The same is true of checking enabled extensions in the Aggregates panel.

The desired experience is the user should still be able to log in and
perform actions unrelated to those misconfigured/down services.

Closes-Bug: #1406329
Change-Id: I8b0eae53c1b8c9a8bc9376e6d813dd6d2bea0e9b